requirements.txt
asgiref==3.8.1
awsebcli==3.20.10
blessed==1.20.0
cachetools==5.4.0
cement==2.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.3
Django==4.2.15
docutils==0.16
google-api-core==2.19.1
google-api-python-client==2.141.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.8
jmespath==1.0.1
mysqlclient==2.2.4
oauthlib==3.2.2
pathspec==0.10.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.7.2
botocore==1.31.85  # awsebcliが要求するバージョン
s3transfer==0.4.2  # botocore==1.31.85に適合するバージョン
semantic-version==2.8.5
setuptools==73.0.1
six==1.16.0
sqlparse==0.5.1
stripe==10.8.0
termcolor==1.1.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.19
wcwidth==0.1.9